api-gateway/requirements.txt
django
djangorestframework
pyjwt
requests
mysqlclient